feat: add button to zoom in selected roi
